shifting issue fixed

+++ hex.cpp
@@ -10,6 +10,7 @@ std::string hex8(uint8_t i)
     // static_cast is necessary to prevent the insertion operator << from treating the 8-bit
     // integer as a character and printing it incorrectly
     os << std::hex << std::setfill('0') << std::setw(2) << static_cast<uint16_t>(i);
+    // os << std::hex << std::setfill('0') << std::setw(2) << static_cast<uint16_t>(i);
     return os.str();
 }
 
@@ -17,7 +18,8 @@ std::string hex8(uint8_t i)
 std::string hex32(uint32_t i)
 {
     std::ostringstream os;
-    os << std::hex << std::setfill('0') << std::setw(8) << i;
+    os << std::hex << std::setfill('0') << std::setw(8) << static_cast<uint32_t>(i);
+    // os << std::hex << std::setfill('0') << std::setw(8) << i;
     return os.str();
 }
